Made some updates...
The 'isscalar(par)' is just to ensure that the function is not mistakenly passed as an array of structures.  I also uncommented the 'elseif ~ischar(par)...' condition because this allows you to pass a numerical vector directly (without the structure like the first condition).  We don't have to worry about this writing a scalar to a buffer because the 'elseif isscalar(par)' condition precedes it and will catch any of those and use the appropriate SetTagVal command.  Please let me know if I screwed something up or if this doesn't help with your problem.